Upstream removed rows the menus here index by number, which is the one
kind of conflict git cannot see: nothing fails to compile and nothing
looks wrong, the wrong string is simply drawn.

Menu_Letter_Data lost AUDIO from the sound menu, so every entry these
chantiers appended sits one lower -- DISPLAY_LABEL_FIRST, CUSTOM_THEME_
FIRST, BACKGROUND_STAGE_FIRST and the six COL_EDIT_* all move down by
one. Slide_Pos_Data_61 lost the matching position, and the two tables are
verified to be the same length again with each named constant checked
against the string it is supposed to name.

Slide_Pos_Data_64 went the other way: upstream added a position at 18, so
the four *_VALUE_POS_FIRST move up by one. Letter_Data_64 lost the Sound
Mode row, so every LETTER_TYPE_* above it moves down by one. AT_Jmp_Tbl
lost Memory_Card, Save_Direction and Load_Direction, so the six
MENU_SCREEN_* here move down by two, and the option menu now routes
through upstream's option_routines with the DISPLAY row still overridden.

The sound menu's own renumbering follows from the same removal:
Convert_Buff[3][1][n] became [n-1], SOUND_ITEM_BGM_TYPE is 2, and its
EXIT row is 4. col_edit picks up texgroup_data.h, texgrpdat having moved
there in the texgroup refactor.

+++ src/sf33rd/Source/Game/debug/debug_config.h
@@ -16,7 +16,7 @@ typedef struct {
     int stage_override;
     u8 cpu_sa;
     bool turbo_buttons;
-    u8 bonus_stage_override;
+    int bonus_stage_override;
     bool skip_to_ending;
     u8 cpu_active_override;
     u8 cpu_passive_override;
