+++ src/sf33rd/Source/Game/screen/ranking.c
@@ -325,7 +325,7 @@ void Ranking_01_2nd() {
         Setup_Demo_Stage();
         Push_LDREQ_Queue_Player(0, My_char[0]);
         Push_LDREQ_Queue_Player(1, My_char[1]);
-        Push_LDREQ_Queue_BG((s32)bg_w.stage);
+        Push_LDREQ_Queue_BG(bg_w.stage);
     }
 }
